+++ sdk/resourcemanager/computeschedule/armcomputeschedule/models_serde.go
@@ -484,10 +484,10 @@ func (o *OperationDisplay) UnmarshalJSON(data []byte) error {
 // MarshalJSON implements the json.Marshaller interface for type OperationErrorDetails.
 func (o OperationErrorDetails) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
-	populate(objectMap, "crpOperationId", o.CrpOperationID)
+	populate(objectMap, "azureOperationName", o.AzureOperationName)
 	populate(objectMap, "errorCode", o.ErrorCode)
-	populateDateTimeRFC3339(objectMap, "errorDetails", o.ErrorDetails)
-	populateDateTimeRFC3339(objectMap, "timeStamp", o.TimeStamp)
+	populate(objectMap, "errorDetails", o.ErrorDetails)
+	populateDateTimeRFC3339(objectMap, "timestamp", o.Timestamp)
 	return json.Marshal(objectMap)
 }
 
@@ -500,17 +500,17 @@ func (o *OperationErrorDetails) UnmarshalJSON(data []byte) error {
 	for key, val := range rawMsg {
 		var err error
 		switch key {
-		case "crpOperationId":
-			err = unpopulate(val, "CrpOperationID", &o.CrpOperationID)
+		case "azureOperationName":
+			err = unpopulate(val, "AzureOperationName", &o.AzureOperationName)
 			delete(rawMsg, key)
 		case "errorCode":
 			err = unpopulate(val, "ErrorCode", &o.ErrorCode)
 			delete(rawMsg, key)
 		case "errorDetails":
-			err = unpopulateDateTimeRFC3339(val, "ErrorDetails", &o.ErrorDetails)
+			err = unpopulate(val, "ErrorDetails", &o.ErrorDetails)
 			delete(rawMsg, key)
-		case "timeStamp":
-			err = unpopulateDateTimeRFC3339(val, "TimeStamp", &o.TimeStamp)
+		case "timestamp":
+			err = unpopulateDateTimeRFC3339(val, "Timestamp", &o.Timestamp)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -654,7 +654,7 @@ func (r ResourceOperationDetails) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "retryPolicy", r.RetryPolicy)
 	populate(objectMap, "state", r.State)
 	populate(objectMap, "subscriptionId", r.SubscriptionID)
-	populate(objectMap, "timeZone", r.TimeZone)
+	populate(objectMap, "timezone", r.Timezone)
 	return json.Marshal(objectMap)
 }
 
@@ -697,8 +697,8 @@ func (r *ResourceOperationDetails) UnmarshalJSON(data []byte) error {
 		case "subscriptionId":
 			err = unpopulate(val, "SubscriptionID", &r.SubscriptionID)
 			delete(rawMsg, key)
-		case "timeZone":
-			err = unpopulate(val, "TimeZone", &r.TimeZone)
+		case "timezone":
+			err = unpopulate(val, "Timezone", &r.Timezone)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -800,9 +800,9 @@ func (r *RetryPolicy) UnmarshalJSON(data []byte) error {
 // MarshalJSON implements the json.Marshaller interface for type Schedule.
 func (s Schedule) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
-	populateDateTimeRFC3339(objectMap, "deadLine", s.DeadLine)
+	populateDateTimeRFC3339(objectMap, "deadline", s.Deadline)
 	populate(objectMap, "deadlineType", s.DeadlineType)
-	populate(objectMap, "timeZone", s.TimeZone)
+	populate(objectMap, "timezone", s.Timezone)
 	return json.Marshal(objectMap)
 }
 
@@ -815,14 +815,14 @@ func (s *Schedule) UnmarshalJSON(data []byte) error {
 	for key, val := range rawMsg {
 		var err error
 		switch key {
-		case "deadLine":
-			err = unpopulateDateTimeRFC3339(val, "DeadLine", &s.DeadLine)
+		case "deadline":
+			err = unpopulateDateTimeRFC3339(val, "Deadline", &s.Deadline)
 			delete(rawMsg, key)
 		case "deadlineType":
 			err = unpopulate(val, "DeadlineType", &s.DeadlineType)
 			delete(rawMsg, key)
-		case "timeZone":
-			err = unpopulate(val, "TimeZone", &s.TimeZone)
+		case "timezone":
+			err = unpopulate(val, "Timezone", &s.Timezone)
 			delete(rawMsg, key)
 		}
 		if err != nil {